Switch to vector/matrices to double

+++ src/camera.h
@@ -4,10 +4,10 @@
 class Camera
 {
 public:
-    glm::vec3 eye;
-    glm::vec3 lookAt;
-    glm::vec3 up;
-    glm::vec3 right;
+    glm::dvec3 eye;
+    glm::dvec3 lookAt;
+    glm::dvec3 up;
+    glm::dvec3 right;
     decimal fov; // angle de vue total horizontal, en radians
     decimal imageDist; // distance entre l'oeil et le plan image
